Update patch set 4
Patch Set 4: Code-Review-1

(3 comments)

I think you missed some points to change from databaseType to database and database to databaseConfig.

I think it's OK to use type, but we're changing the meaning of 'database' mid-code and this is kind of confusing.
When we get it from settings, 'database' means the name (or type) of the DB. Afterwards it means the db config (or settings) that we got from the settings file.

Patch-set: 4
Reviewer: Gerrit User 1015148 <1015148@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1