fix: Use GOPROXY to download dependencies
As of Go v1.13, GOPROXY is used by default with 'go' commands. By
default, GOPROXY is set to “https://proxy.golang.org,direct", meaning
it will only use 'direct' (VCS) to download modules if the proxy url is
unable to find them. Once a module is cached and stored in the Go proxy,
it is in immutable storage forever. This means there is no need to use
the '/vendor' directory to provide resiliency in the event of external
sources going down, disappearing, or moving.

From the above, 'go mod vendor' is no longer needed, and neither will
'go mod download' -- it doesn't use GOPROXY and 'go build' itself will
reach out to the network as needed to satisfy imports).

As a result, 'make deps' is not used -- remove it from the 'Makefile'
and 'Dockerfile'.