data: Add domain query entity
In order to reduce the number of query parameters passed to the use
case operations, this patch creates an domain entity to represent them.
This change will improve the code readability and the query handling
itself. In addition, this patch removes unnecessary dot import from the
controller and rename the query extractor method to getURLQueryParams.