data: interactor: Add initial interface
This change includes the initial step to refactor the interactors in a
way that they become readable and easy to change. It adds an interface
to abstract the use case operations and group it in a single file along
with the `DataInteractor` structure and its constructor. The operations
will be incrementally updated to be in accordance with the interface.

Besides, this patch also removes unnecessary dot import from the
interactor.