data: refactor: Enable fine-grained authorization
Previously, the service was allowing to operate on data by using a
valid user authorization token. However, the user should operate only
on their things. In order to achieve this fine-grained authorization,
this patch search for the user's things on the thing's service and
does the necessary validations. The main consequence of this patch is
that it replaces the users service by the things one and import the
babeltower http package that already provides a proxy for that service.