data: Update data handler to listen new event
As the KNot Cloud core API changed recently, data messages are now
coming through `data.published` events, which can be received via
fanout exchange with the same name. In order to support this new
scenario, this patch updates the amqp and handler operations to enable
receiving broadcasted messages. Also, it validates if authorization
token is correctly provided and log when data is successfully saved.